+++ .devcontainer/devcontainer.json
@@ -51,7 +51,6 @@
 		"ms-azuretools.vscode-docker",
 		"streetsidesoftware.code-spell-checker",
 		"oderwat.indent-rainbow",
-		"littlefoxteam.vscode-python-test-adapter",
 		"GitHub.vscode-pull-request-github"
 	],
 	"forwardPorts": [
