+++ conda/environment.yml
@@ -13,11 +13,9 @@ dependencies:
       - ipykernel
       - nbconvert
       - nbformat
-      - rospkg
+      - rosbags
       - kaleido
       - pyntcloud
-      - py3rosmsgs
-      - pycryptodomex
       - setuptools
       - blackcellmagic
       - isort
