--- docker/python_logging.conf
@@ -1,53 +1,0 @@
-[loggers]
-keys=root,fileLogger,rotatingFileLogger
-
-[handlers]
-keys=consoleHandler,fileHandler,rotatingFileHandler
-
-[formatters]
-keys=simpleFormatter
-
-[logger_root]
-level=DEBUG
-handlers=consoleHandler
-
-[logger_fileLogger]
-level=DEBUG
-# The logger configured handler
-handlers=fileHandler
- # Logger's name
-qualname=fileLogger
-propagate=0
-
-[logger_rotatingFileLogger]
-level=DEBUG
- # This configuration, rotatingFileLogger in to configure the consoleHandler, rotatingFileHandler
- # ConsoleHandler responsible for the log output to the console
- # RotatingFileHandler responsible to save the log output to a file
-handlers=consoleHandler,rotatingFileHandler
-qualname=rotatingFileLogger
-propagate=0
-
-[handler_consoleHandler]
-class=FileHandler
-level=DEBUG
-formatter=simpleFormatter
-#args=(sys.stdout,)
-args=("/tmp/stdout.log",)
-
-[handler_fileHandler]
-class=FileHandler
-level=DEBUG
-formatter=simpleFormatter
-args=('/tmp/logging.log', 'a')
-
-[handler_rotatingFileHandler]
-class=handlers.RotatingFileHandler
-level=WARNING
-formatter=simpleFormatter
-args=("/tmp/rotating_logging.log", "a", 1*1024*1024, 5)
-
-[formatter_simpleFormatter]
-#format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
-format=%(asctime)s - %(module)s - %(thread)d - %(levelname)s : %(message)s
-datefmt=%Y-%m-%d %H:%M:%S
