+++ pointcloudset/io/__init__.py
@@ -2,10 +2,11 @@
 Functions for file input and output.
 """
 
-from pyntcloud.io import \
-    FROM_FILE as POINTCLOUD_FROM_FILE  # needs to be here, ignore warnings
+from pyntcloud.io import (
+    FROM_FILE as POINTCLOUD_FROM_FILE,
+)  # needs to be here, ignore warnings
 
-from pointcloudset.io.dataset.bag import dataset_from_rosbag
+from pointcloudset.io.dataset.rosbag import dataset_from_rosbag
 from pointcloudset.io.dataset.dir import dataset_from_dir, dataset_to_dir
 from pointcloudset.io.dataset.pointcloud import dataset_from_pointclouds
 from pointcloudset.io.pointcloud.csv import write_csv
